version 1.2.1
Improved block_collision: the function now takes into account situations where the character goes beyond the top of the screen and handles the collision outcomes accordingly. If a structure exists at the highest level of the map frame & directly after the character's position, the character cannot jump on top of it.

Added 'grounded' feature in physics_t: when a physics_t variable is 'grounded', it avoids holes and pits by counting them as collisions.