version 1.2.2
Added sprites for the Moria world
Changed the x-axis reference point of the character from the left of the display to the origin of the map (map_x = 0). This improves the clarity and robustness of the code by aligning all physics_t elements to a common origin.
Added platforms feature to the game.